insurance_genai_bot/config/requirements.txt
python-dotenv
scikit-learn
streamlit
openai
langchain
langchain-community
langchain-openai
unstructured
networkx
openpyxl
python-docx
msoffcrypto-tool